实现日志的UI分离
