+++ src/app/widgets/StatusPanel/StatusPanel.py
